[S07P31C209-233][BE-정지원] chore : likeCnt 타입변경

+++ msa_gradle/img/src/main/java/com/fairytail/img/dto/PostDto.java
@@ -25,5 +25,5 @@ public class PostDto {
     private LocalDateTime date;
     private Integer status;
     private MultipartFile file;
-    private Long likeCnt;
+    private Integer likeCnt;
 }
